boards: st: h7: add memory-regions property to mac node
Use the memory-regions property in DT to explicitly locate ethernet
buffer and descriptor to an sram region. Since sram3 was the first
option and sram2 the fallback (in mpu_regions.c and sections.ld),
try to use sram3 when possible; otherwise use sram2.

Add #memory-region-cells to sram3 and sram2 nodes in dts/arm/st/h7
since it is needed by the memory-regions property.

Use the same check and definition of the sram_eth_node in mpu_regions.c
and sections.ld as in soc/st/stm32/h7rsx.

Signed-off-by: Benjamin Klaric <benjamin.klaric01@gmail.com>